Improved configuration.
