+++ stream-compositions/services/legal-entity-composition-service/src/main/java/com/backbase/stream/compositions/legalentity/core/config/LegalEntityConfiguration.java
@@ -1,16 +1,44 @@
 package com.backbase.stream.compositions.legalentity.core.config;
 
+import com.backbase.stream.compositions.integration.legalentity.ApiClient;
+import com.backbase.stream.compositions.integration.legalentity.api.LegalEntityIntegrationApi;
 import com.backbase.stream.compositions.legalentity.core.mapper.LegalEntityMapper;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import lombok.AllArgsConstructor;
 import org.mapstruct.factory.Mappers;
-import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Primary;
+import org.springframework.web.reactive.function.client.WebClient;
+
+import java.text.DateFormat;
 
 @Configuration
-@ConfigurationProperties(prefix = "backbase.stream.compositions.legal-entity")
+@AllArgsConstructor
+@EnableConfigurationProperties(LegalEntityConfigurationProperties.class)
 public class LegalEntityConfiguration {
+    private final LegalEntityConfigurationProperties legalEntityConfigurationProperties;
+
     @Bean
     public LegalEntityMapper mapper() {
         return Mappers.getMapper(LegalEntityMapper.class);
+    }
+
+    @Bean
+    @Primary
+    public LegalEntityIntegrationApi legalEntityIntegrationApi(ApiClient legalEntityClient) {
+        return new LegalEntityIntegrationApi(legalEntityClient);
+    }
+
+    @Bean
+    public ApiClient legalEntityClient(
+            WebClient dbsWebClient,
+            ObjectMapper objectMapper,
+            DateFormat dateFormat) {
+        ApiClient apiClient = new ApiClient(dbsWebClient, objectMapper, dateFormat);
+        apiClient.setBasePath(legalEntityConfigurationProperties.getLegalEntityIntegrationUrl());
+
+        return apiClient;
     }
 }
